+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -62,6 +62,7 @@ import at.hannibal2.skyhanni.features.nether.ashfang.*
 import at.hannibal2.skyhanni.features.nether.reputationhelper.CrimsonIsleReputationHelper
 import at.hannibal2.skyhanni.features.rift.RiftTimer
 import at.hannibal2.skyhanni.features.rift.ShowMotesNpcSellPrice
+import at.hannibal2.skyhanni.features.rift.VampireSlayerFeatures
 import at.hannibal2.skyhanni.features.rift.area.RiftLarva
 import at.hannibal2.skyhanni.features.rift.area.colosseum.BlobbercystsHighlight
 import at.hannibal2.skyhanni.features.rift.area.dreadfarm.RiftAgaricusCap
@@ -340,6 +341,7 @@ class SkyHanniMod {
         loadModule(HighlightMiningCommissionMobs())
         loadModule(ShowMotesNpcSellPrice())
         loadModule(LivingMetalSuitProgress())
+        loadModule(VampireSlayerFeatures())
         loadModule(BlobbercystsHighlight())
 
         init()