+++ src/main/java/at/hannibal2/skyhanni/config/features/RiftConfig.java
@@ -447,6 +447,7 @@ public class RiftConfig {
 
         public static class LivingCaveConfig {
 
+
             @Expose
             @ConfigOption(name = "Living Metal Suit Progress", desc = "")
             @Accordion